adds maxTries param to request
