+++ pkg/workflows/sdk/runtime.go
@@ -30,11 +30,12 @@ type Runtime interface {
 }
 
 type FetchRequest struct {
-	URL       string            `json:"url"`                 // URL to query, only http and https protocols are supported.
-	Method    string            `json:"method,omitempty"`    // HTTP verb, defaults to GET.
-	Headers   map[string]string `json:"headers,omitempty"`   // HTTP headers, defaults to empty.
-	Body      []byte            `json:"body,omitempty"`      // HTTP request body
-	TimeoutMs uint32            `json:"timeoutMs,omitempty"` // Timeout in milliseconds
+	URL        string            `json:"url"`                 // URL to query, only http and https protocols are supported.
+	Method     string            `json:"method,omitempty"`    // HTTP verb, defaults to GET.
+	Headers    map[string]string `json:"headers,omitempty"`   // HTTP headers, defaults to empty.
+	Body       []byte            `json:"body,omitempty"`      // HTTP request body
+	TimeoutMs  uint32            `json:"timeoutMs,omitempty"` // Timeout in milliseconds is the max time elapsed for a request including any retries.
+	MaxRetries uint32            `json:"maxTries,omitempty"`  // Max number of times to retry a request.  A request is made once, but a default value of 0 implies no retries.
 }
 
 type FetchResponse struct {
