+++ pkg/workflows/wasm/host/module.go
@@ -478,11 +478,12 @@ func toSdkReq(req *wasmpb.FetchRequest) *FetchRequest {
 	}
 	return &FetchRequest{
 		FetchRequest: sdk.FetchRequest{
-			URL:       req.Url,
-			Method:    req.Method,
-			Headers:   h,
-			Body:      req.Body,
-			TimeoutMs: req.TimeoutMs,
+			URL:        req.Url,
+			Method:     req.Method,
+			Headers:    h,
+			Body:       req.Body,
+			TimeoutMs:  req.TimeoutMs,
+			MaxRetries: req.MaxRetries,
 		},
 		Metadata: md,
 	}
